server/leadscore-api-py/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asn1crypto==1.5.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloud-sql-python-connector==1.9.0
colorama==0.4.6
cryptography==46.0.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.115.0
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-error-reporting==1.9.1
google-cloud-logging==3.9.0
google-genai==0.3.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpcio==1.75.0
grpcio-status==1.62.3
h11==0.16.0
httptools==0.6.4
idna==3.10
multidict==6.6.4
pg8000==1.31.2
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==2.0.7
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
scramp==1.4.6
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.38.6
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.30.1
watchfiles==1.1.0
websockets==14.2
wheel==0.45.1
yarl==1.20.1
fastapi
uvicorn[standard]
google-genai
cloud-sql-python-connector[pg8000]
pg8000
sqlalchemy
pydantic
python-json-logger
google-cloud-logging
google-cloud-error-reporting
passlib[bcrypt]
python-dotenv
email-validator
